+++ libs/search-widget/src/old-components/row/Square.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
-  import type { IResource } from '@nuclia/core';
+  import { IResource, ReadableResource } from '@nuclia/core';
   import { formatDate, formatTitle } from '../../core/utils';
   import Thumbnail from '../../common/thumbnail/Thumbnail.svelte';
   import { goToResource } from '../results/results.utils';
 
   export let result: IResource;
   let labels: string[];
-  $: labels = (result.usermetadata?.classifications || []).map((label) => label.label);
+  $: labels = new ReadableResource(result).getClassifications().map((label) => label.label);
 </script>
 
 <div