+++ libs/search-widget/src/old-components/viewer/Metadata.svelte
@@ -8,7 +8,15 @@
   import Button from '../../common/button/Button.svelte';
   import { onDestroy } from 'svelte';
   import { annotationMode, annotations, selectedFamily } from '../../core/stores/annotation.store';
-  import { files, links, previewLinks, resource, summaries } from '../../core/stores/resource.store';
+  import {
+    files,
+    links,
+    previewLinks,
+    resource,
+    summaries,
+    resourceHasEntities,
+    resourceLabels,
+  } from '../../core/stores/resource.store';
   import { entityGroups } from '../../core/stores/entities.store';
   import Label from '../../common/label/Label.svelte';
   import { searchBy } from '../../common/label/label.utils';
@@ -101,7 +109,7 @@
   <div class="entities">
     <Entities />
   </div>
-  {#if !$annotationMode}
+  {#if !$annotationMode && $resourceHasEntities}
     <div class="entities">
       <h3>{$_('entities.annotated')}</h3>
       <Entities showAnnotated={true} />
@@ -144,11 +152,11 @@
         </div>
       {/if}
 
-      {#if ($resource.usermetadata?.classifications || []).length > 0}
+      {#if $resourceLabels.length > 0}
         <div class="metadata-value">
           <h3>{$_('resource.classification')}</h3>
           <div class="labels">
-            {#each $resource.usermetadata?.classifications || [] as label}
+            {#each $resourceLabels as label}
               <Label
                 {label}
                 clickable={$widgetType === 'search'}